requirements.txt
streamlit
groq
python-dotenv
fpdf
